agents/congress_trades_agent/requirements.txt
google-adk
db-dtypes
pydantic
pandas
pyarrow
google-cloud-aiplatform
google-cloud-bigquery
yfinance
requests
db-dtypes
rapidfuzz 